moved hash-related stuff core.digest package. Renamed XorSeaHash to BalticHash :).
