+++ src/main/java/net/szum123321/textile_backup/core/create/FileInputStreamSupplier.java
@@ -20,7 +20,8 @@ package net.szum123321.textile_backup.core.create;
 
 import net.szum123321.textile_backup.TextileBackup;
 import net.szum123321.textile_backup.TextileLogger;
-import net.szum123321.textile_backup.core.FileTreeHashBuilder;
+import net.szum123321.textile_backup.core.digest.FileTreeHashBuilder;
+import net.szum123321.textile_backup.core.digest.HashingInputStream;
 
 import java.io.IOException;
 import java.io.InputStream;
